feature1_chatbot/requirements.txt
scikit-learn
numpy
sentence-transformers
faiss-cpu
tqdm
